// src/utils/keycap-type.ts
const enum KeycapType {
    Esc = 'Esc',
    F1 = 'F1',
    F2 = 'F2',
    F3 = 'F3',
    F4 = 'F4',
    F5 = 'F5',
    F6 = 'F6',
    F7 = 'F7',
    F8 = 'F8',
    F9 = 'F9',
    F10 = 'F10',
    F11 = 'F11',
    F12 = 'F12',
    Apostrophe = 'Apostrophe',
    _1 = '_1',
    _2 = '_2',
    _3 = '_3',
    _4 = '_4',
    _5 = '_5',
    _6 = '_6',
    _7 = '_7',
    _8 = '_8',
    _9 = '_9',
    _0 = '_0',
    Hyphen = 'Hyphen',
    Plus = 'Plus',
    Backspace = 'Backspace',
    Tab = 'Tab',
    Q = 'Q',
    W = 'W',
    E = 'E',
    R = 'R',
    T = 'T',
    Y = 'Y',
    U = 'U',
    I = 'I',
    O = 'O',
    P = 'P',
    OpenBracket = 'OpenBracket',
    CloseBracket = 'CloseBracket',
    ReverseSlash = 'ReverseSlash',
    CapsLock = 'CapsLock',
    A = 'A',
    S = 'S',
    D = 'D',
    F = 'F',
    G = 'G',
    H = 'H',
    J = 'J',
    K = 'K',
    L = 'L',
    SemiColon = 'SemiColon',
    QuotationMark = 'QuotationMark',
    Enter = 'Enter',
    LShift = 'LShift',
    Z = 'Z',
    X = 'X',
    C = 'C',
    V = 'V',
    B = 'B',
    N = 'N',
    M = 'M',
    Comma = 'Comma',
    Period = 'Period',
    Slash = 'Slash',
    RShift = 'RShift',
    LCtrl = 'LCtrl',
    LAlt = 'LAlt',
    LCommand = 'LCommand',
    Space = 'Space',
    RCommand = 'RCommand',
    RAlt = 'RAlt',
    RCtrl = 'RCtrl',
}


export default KeycapType